+++ CHANGELOG.md
@@ -71,7 +71,6 @@
 - add const keys for RocketController [rocketRequestKey], [sizeScreenKey] & [sizeDesignKey]
 
 ## [1.0.1]
-- Use [enums] insread of [string] for http methods in [sendFile] method on [RocketRequest].
+- Use [enums] instead of [string] for http methods in [sendFile] method on [RocketRequest].
 = Add [RocketState] enums for manage model states (done,loading,failed).
-- Add [enableDebug] on Models for get model state & duration of loading
-
+- Add [enableDebug] on Models for log model state & duration of loading