+++ CHANGELOG.md
@@ -70,7 +70,7 @@
 - use [LinkedList] instead of [List] & add extensions needed
 - add const keys for RocketController [rocketRequestKey], [sizeScreenKey] & [sizeDesignKey]
 
-## [1.0.1]
+## [1.0.2]
 - Use [enums] instead of [string] for http methods in [sendFile] method on [RocketRequest].
 - Add [RocketState] enums for manage model states (done,loading,failed).
 - Add [enableDebug] on Models for log model state & duration of loading